+++ changelog.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Convex.MockChain`: Export `fromLedgerUTxO`
   - `Convex.MockChain`: `MonadTrans` instance for `MockchainT`, export constructor
   - `Convex.MockChain`: Add `MockchainIO` type
+  - `Convex.MockChain`: Add `resolveDatumHash` function, giving access to all datums that were previously seen
 * `convex-node-client`: Add `Convex.NodeClient.WaitForTxnClient`
 * `Convex.BuildTx`:
   - Add a monadic (writer) interface for building transactions